+++ src/main/java/picocli/CommandLine.java
@@ -5612,6 +5612,10 @@ public class CommandLine {
     private static class DefaultFactory implements IFactory {
         static Class<?> GROOVY_CLOSURE_CLASS = loadClosureClass();
         private static Class<?> loadClosureClass() {
+            if (Boolean.getBoolean("picocli.disable.closures")) {
+                tracer().info("DefaultFactory: groovy Closures in annotations are disabled and will not be loaded");
+                return null;
+            }
             try { return Class.forName("groovy.lang.Closure"); }
             catch (Exception ignored) { return null;}
         }